Break up embedded graphics tests
