+++ tests/embedded_graphics_target/assets/rhombic-dodecahedron.tga
