+++ tests/mod.rs
@@ -1,4 +1,6 @@
 #[macro_use]
 mod common;
+#[cfg(feature = "embedded-graphics")]
+mod embedded_graphics_target;
 mod scroll_view;
 mod transition;
